requirements.txt
pytoniq
pyraptorq
aiohttp
python-dotenv
typing_extensions
asyncpg
bitstring
beautifulsoup4
